New Tests and BigDecimal

+++ src/main/java/com/andersentask/bookshop/book/entities/Book.java
@@ -6,16 +6,19 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class Book {
+
     private Long id;
 
     private String name;
 
     private BookStatus status;
 
-    private Double price;
+    private BigDecimal price;
 }
